Restricting users and roles fetch permission + public view create & update

+++ src/modules/user/user.service.ts
@@ -5,6 +5,7 @@ import { UserRepository } from './providers/user.repository';
 import { ICount, IResponse } from 'src/core/interfaces/response.interface';
 import { DeleteUsersDto } from './dto/delete-user.dto';
 import { AuthHelper } from 'src/core/helpers/auth.helper';
+import { ENUM_ROLE_TYPE } from 'src/core/constants/role.constants';
 
 @Injectable()
 export class UserService {
@@ -13,9 +14,27 @@ export class UserService {
     private readonly authHelper: AuthHelper,
   ) {}
 
-  async findAll(): Promise<IResponse<User[]>> {
+  async findAll(req: any): Promise<IResponse<User[]>> {
     try {
-      const users = await this.userRepo.findAll();
+      const excludedRoles = [];
+
+      const loggedInUser = await this.findById(req?.user?.id);
+      if (!loggedInUser) {
+        throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+      }
+      const userRoles = loggedInUser.data.roles.map((role) => role.type);
+      if (userRoles.includes(ENUM_ROLE_TYPE.SUPER_ADMIN)) {
+        excludedRoles.push(ENUM_ROLE_TYPE.SUPER_ADMIN);
+      } else if (userRoles.includes(ENUM_ROLE_TYPE.ADMIN)) {
+        excludedRoles.push(ENUM_ROLE_TYPE.SUPER_ADMIN, ENUM_ROLE_TYPE.ADMIN);
+      } else if (userRoles.includes(ENUM_ROLE_TYPE.MANAGER)) {
+        excludedRoles.push(
+          ENUM_ROLE_TYPE.SUPER_ADMIN,
+          ENUM_ROLE_TYPE.ADMIN,
+          ENUM_ROLE_TYPE.MANAGER,
+        );
+      }
+      const users = await this.userRepo.findAll(excludedRoles);
       return {
         statusCode: HttpStatus.OK,
         message: 'Users retrieved successfully',
@@ -86,6 +105,9 @@ export class UserService {
   ): Promise<IResponse<User>> {
     try {
       const {} = updateUserDto;
+      const user = await this.userRepo.findById(id);
+      if (!user)
+        throw new HttpException('User not found', HttpStatus.NOT_FOUND);
       const newUser = await this.userRepo.update(id, updateUserDto);
       return {
         statusCode: HttpStatus.OK,
@@ -100,8 +122,32 @@ export class UserService {
     }
   }
 
+  async updatePublicDisplay(id: string): Promise<IResponse<User>> {
+    try {
+      const user = await this.userRepo.findById(id);
+      if (!user)
+        throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+      const newUser = await this.userRepo.update(id, {
+        publicDisplay: !user.publicDisplay,
+      });
+      return {
+        statusCode: HttpStatus.OK,
+        message: `User ${newUser[1][0].publicDisplay ? 'set to ' : 'removed from '} public display`,
+        data: newUser[1][0],
+      };
+    } catch (error) {
+      throw new HttpException(
+        error.message || 'Server Error',
+        error.status || HttpStatus.INTERNAL_SERVER_ERROR,
+      );
+    }
+  }
+
   async deleteOne(id: string): Promise<IResponse<ICount>> {
     try {
+      const user = await this.userRepo.findById(id);
+      if (!user)
+        throw new HttpException('User not found', HttpStatus.NOT_FOUND);
       const count = await this.userRepo.deleteOne(id);
       return {
         statusCode: HttpStatus.OK,
